Update: files position refactored
